requirements.txt
requests==2.31.0
PyYAML==6.0.1
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
rapidfuzz==3.6.1 